feat(ta_codegen): expand YAML IDL schema with full metadata from C abstract tables
- flags fields are now lists (e.g., [overlap, unstable_period])
- Added unstable_period flag to RSI and EMA
- Added suggested optimization hints to optional inputs
- default changed to f64 to support real-valued defaults
- Output flags now a list to support multiple display hints
- Added IDL reference docs at docs/ta_func_defs_idl.md
- All 48 tests + 36 validations passing

+++ tools/ta_codegen/src/backends/rust_lang.rs
@@ -112,7 +112,7 @@ fn gen_lookback(func: &FuncDef, snake: &str, enums: &HashMap<String, EnumDef>) -
             if let (Some(default), Some((lo, hi))) = (opt.default, opt.range) {
                 out.push_str(&format!(
                     "    /// * `{}` - Number of period (default: {}, range: {}..={})\n",
-                    opt.name, default, lo, hi
+                    opt.name, default as i64, lo, hi
                 ));
             }
         }
@@ -211,7 +211,7 @@ fn gen_public_func(func: &FuncDef, snake: &str, single_precision: bool) -> Strin
             if let (Some(default), Some((lo, hi))) = (opt.default, opt.range) {
                 out.push_str(&format!(
                     "    /// * `{}` - Number of period (default: {}, range: {}..={})\n",
-                    opt.name, default, lo, hi
+                    opt.name, default as i64, lo, hi
                 ));
             }
         }
@@ -643,7 +643,7 @@ fn gen_opt_param_validation(opt: &OptInput, pad: &str, is_lookback: bool) -> Str
             ));
             out.push_str(&format!(
                 "{}    {} = {};\n",
-                pad, name, default
+                pad, name, default as i64
             ));
 
             if let Some((lo, hi)) = opt.range {
